Add SerialInterface and add exists storage function

+++ src/systems/Settings.h
@@ -9,7 +9,7 @@ public:
   Settings(StorageInterface &stg);
 
   uint8_t timezone = 7;
-  uint8_t updateInterval_ms = 2000;
+  uint32_t updateInterval_ms = 2000;
 
   bool load(std::string filename = "/settings.json");
   bool store(std::string filename = "/settings.json");
